requirements.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.6.1
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.18.4
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.1
psycopg2==2.9.10
py4j==0.10.9.7
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
schedule==1.2.2
selenium==4.27.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
spark-nlp==5.5.2
streamlit==1.41.1
streamlit-autorefresh==1.0.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
